apex-v4.2.0/pyproject.toml
[tool.poetry]
name = "apex"
version = "4.2.0"
description = "Middleware implementation for SAPIENT"
authors = ["Roke"]
readme = "readme.md"
packages = [
  { include = "sapient_apex_api" },
  { include = "sapient_apex_gui" },
  { include = "sapient_apex_qt_helpers" },
  { include = "sapient_apex_replay" },
  { include = "sapient_apex_server" },
  { include = "sapient_msg" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
protobuf = "^4.25.1"
trio = "^0.23.1"
ulid-py = "^1.1.0"
fastapi = "^0.104.1"
pyside6 = { version = "6.5.1.1", optional = true }
uvicorn = "^0.24.0.post1"
httpx = "^0.26.0"
elasticsearch = "^8.11.1"
sqlalchemy = "^2.0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-trio = "^0.8.0"
black = "^23.11.0"
nuitka = "^1.9.3"
grpcio-tools = "^1.59.3"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"
poetry-plugin-export = "^1.6.0"
python-dotenv = "^1.0.0"
gitpython = "^3.1.41"
click = "^8.1.7"
plotille = "^5.0.0"

[tool.poetry.extras]
gui = ["pyside6"]

[tool.poetry.scripts]
apex = { reference = "sapient_apex_server:apex.serve_apex", type = "console" }
apex_gui = { reference = "sapient_apex_gui:apex_gui.main", type = "console" }
apex_replay = { reference = "sapient_apex_replay:replay.main", type = "console" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

log_file = "tests/pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

trio_mode = true
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py39']
